+++ src/main/java/dev/moneet/schema/graph/DfsTraversalStrategy.java
@@ -8,14 +8,21 @@ public final class DfsTraversalStrategy implements TraversalStrategy {
     @Override
     public Set<String> traverse(String start, int depth, SchemaGraph graph) {
         Set<String> visited = new LinkedHashSet<>();
-        dfs(start, depth, graph, visited);
+        dfs(start, depth, graph, visited,TraversalDirection.BIDIRECTIONAL);
+        return visited;    }
+
+    @Override
+    public Set<String> traverse(String start, int depth, SchemaGraph graph, TraversalDirection direction) {
+        Set<String> visited = new LinkedHashSet<>();
+        dfs(start, depth, graph, visited,direction);
         return visited;
     }
 
     private void dfs(String current,
                      int depth,
                      SchemaGraph graph,
-                     Set<String> visited) {
+                     Set<String> visited,
+                     TraversalDirection direction) {
 
         if (depth < 0 || visited.contains(current)) {
             return;
@@ -23,12 +30,20 @@ public final class DfsTraversalStrategy implements TraversalStrategy {
 
         visited.add(current);
 
-        for (String neighbor : graph.getDependencies(current)) {
-            dfs(neighbor, depth - 1, graph, visited);
+        if (direction == TraversalDirection.DEPENDENCIES_ONLY
+                || direction == TraversalDirection.BIDIRECTIONAL) {
+
+            for (String neighbor : graph.getDependencies(current)) {
+                dfs(neighbor, depth - 1, graph, visited, direction);
+            }
         }
 
-        for (String neighbor : graph.getDependents(current)) {
-            dfs(neighbor, depth - 1, graph, visited);
+        if (direction == TraversalDirection.DEPENDENTS_ONLY
+                || direction == TraversalDirection.BIDIRECTIONAL) {
+
+            for (String neighbor : graph.getDependents(current)) {
+                dfs(neighbor, depth - 1, graph, visited, direction);
+            }
         }
     }
 }