+++ src/main/java/dev/moneet/schema/graph/TraversalDirection.java
@@ -1,2 +1,7 @@
-package dev.moneet.schema.graph;public enum TraversalDirection {
+package dev.moneet.schema.graph;
+
+public enum TraversalDirection {
+    DEPENDENCIES_ONLY,
+    DEPENDENTS_ONLY,
+    BIDIRECTIONAL
 }